additional RPM dependencies, add config.yml, use RPM files syntax
Included more RPM dependencies that were neglected earlier.

Shipping config.yml as a %config file now. This will not be overwritten
if found to be modified on RPM update. config.yml.rpmsave will be left
on the system if modified on RPM removal.

Explicitly setting %files as opposed to python INSTALLED_FILES methodology.

In order to have config.yml installed via "setup.py install" step it was
added to data_files keyword in setup.py.